fixed a bunch
