+++ src/components/screens/home/screens/Home.js
@@ -1,16 +1,16 @@
 import { ScrollView, Text, View, StyleSheet } from "react-native";
 import React from "react";
-import { color, getColor } from "../../../../assets/colors/color";
+import { color, getColor } from "../../../../../assets/colors/color";
 
-const Tab2 = () => {
+const Home = () => {
   return (
     <View style={styles.container}>
-      <Text style={styles.title}>Tab 2</Text>
+      <Text style={styles.title}>Home</Text>
     </View>
   )
 }
 
-export default Tab2;
+export default Home;
 
 const styles = StyleSheet.create({
   container: {
